refactor: using db for operator form schema
