+++ app/pages/cif/operator/form/[form].tsx
@@ -1,13 +1,30 @@
 import OperatorForm from "components/Operator/OperatorForm";
-import relayFormPageFactory from "lib/pages/relayFormPageFactory";
 import withRelayOptions from "lib/relay/withRelayOptions";
-import { getOperatorsPageRoute } from "routes/pageRoutes";
-import { withRelay } from "relay-nextjs";
+import { RelayProps, withRelay } from "relay-nextjs";
+import { graphql, usePreloadedQuery } from "react-relay/hooks";
+import DefaultLayout from "components/Layout/DefaultLayout";
+import { FormOperatorFormQuery } from "__generated__/FormOperatorFormQuery.graphql";
 
-const [FormPage, query] = relayFormPageFactory(
-  "Operator",
-  getOperatorsPageRoute(),
-  OperatorForm
-);
+const pageQuery = graphql`
+  query FormOperatorFormQuery($form: ID!) {
+    session {
+      ...DefaultLayout_session
+    }
+    formChange(id: $form) {
+      ...OperatorForm_formChange
+    }
+  }
+`;
 
-export default withRelay(FormPage, query, withRelayOptions);
+export function OperatorFormPage({
+  preloadedQuery,
+}: RelayProps<{}, FormOperatorFormQuery>) {
+  const { session, formChange } = usePreloadedQuery(pageQuery, preloadedQuery);
+  return (
+    <DefaultLayout session={session}>
+      <OperatorForm formChange={formChange} />
+    </DefaultLayout>
+  );
+}
+
+export default withRelay(OperatorFormPage, pageQuery, withRelayOptions);
